refactor(tss-2.1): extract createTaskStoreClient() to check-helpers.ts
The URL-check + client-construction pattern (read SHIPWRIGHT_TASK_STORE_URL
and SHIPWRIGHT_TASK_STORE_TOKEN → exit on missing → build fetch client) was
duplicated in check-deploy.ts and check-dev-task.ts. Extract it to a shared
createTaskStoreClient() in check-helpers.ts so all precheck scripts have a
single implementation to maintain.

Addresses review thread on PR #588 (suggestion by @dodizzle).
