+++ plugins/shipwright/scripts/check-deploy.ts
@@ -15,6 +15,7 @@
  */
 
 import {
+  createTaskStoreClient,
   getCurrentUser,
   ghJson,
   readAllowSelfReview,
@@ -246,44 +247,7 @@ interface GhWorkflowRunsJson {
   }>;
 }
 
-function makeTaskStoreFetch() {
-  const taskStoreUrl = (process.env.SHIPWRIGHT_TASK_STORE_URL ?? "").trim();
-  const taskStoreToken = (process.env.SHIPWRIGHT_TASK_STORE_TOKEN ?? "").trim();
-  if (!taskStoreUrl) {
-    process.stderr.write("error: SHIPWRIGHT_TASK_STORE_URL is required\n");
-    process.exit(1);
-  }
-  if (!taskStoreToken) {
-    process.stderr.write("error: SHIPWRIGHT_TASK_STORE_TOKEN is required\n");
-    process.exit(1);
-  }
-  const baseUrl = taskStoreUrl.replace(/\/$/, "");
-  const authHeaders = {
-    Authorization: `Bearer ${taskStoreToken}`,
-    "Content-Type": "application/json",
-  };
-
-  return {
-    async query(params: URLSearchParams): Promise<Task[]> {
-      const res = await fetch(`${baseUrl}/tasks?${params}`, {
-        headers: authHeaders,
-      });
-      if (!res.ok)
-        throw new Error(`task-store GET /tasks?${params} → ${res.status}`);
-      return res.json() as Promise<Task[]>;
-    },
-    async update(id: string, fields: Record<string, unknown>): Promise<Task> {
-      const res = await fetch(`${baseUrl}/tasks/${id}`, {
-        method: "PATCH",
-        headers: authHeaders,
-        body: JSON.stringify(fields),
-      });
-      if (!res.ok)
-        throw new Error(`task-store PATCH /tasks/${id} → ${res.status}`);
-      return res.json() as Promise<Task>;
-    },
-  };
-}
+const makeTaskStoreFetch = createTaskStoreClient;
 
 async function buildProductionDeps(): Promise<Deps> {
   const workspacePath = resolveWorkspacePath();
