+++ plugins/shipwright/scripts/check-helpers.ts
@@ -308,6 +308,56 @@ export async function isMergeOnlyUpdate(
   }
 }
 
+// ─── Task store HTTP client ───────────────────────────────────────────────────
+
+/**
+ * Reads SHIPWRIGHT_TASK_STORE_URL and SHIPWRIGHT_TASK_STORE_TOKEN from the
+ * environment, validates they are present, and returns a minimal fetch client
+ * for the task-store HTTP API.
+ *
+ * Exits with code 1 if either variable is missing so callers (precheck scripts)
+ * get a clean error rather than a confusing undefined/TypeError at call-time.
+ */
+export function createTaskStoreClient(): {
+  query(params: URLSearchParams): Promise<Task[]>;
+  update(id: string, fields: Record<string, unknown>): Promise<Task>;
+} {
+  const taskStoreUrl = (process.env.SHIPWRIGHT_TASK_STORE_URL ?? "").trim();
+  const taskStoreToken = (process.env.SHIPWRIGHT_TASK_STORE_TOKEN ?? "").trim();
+  if (!taskStoreUrl) {
+    process.stderr.write("error: SHIPWRIGHT_TASK_STORE_URL is required\n");
+    process.exit(1);
+  }
+  if (!taskStoreToken) {
+    process.stderr.write("error: SHIPWRIGHT_TASK_STORE_TOKEN is required\n");
+    process.exit(1);
+  }
+  const baseUrl = taskStoreUrl.replace(/\/$/, "");
+  const headers = {
+    Authorization: `Bearer ${taskStoreToken}`,
+    "Content-Type": "application/json",
+  };
+
+  return {
+    async query(params: URLSearchParams): Promise<Task[]> {
+      const res = await fetch(`${baseUrl}/tasks?${params}`, { headers });
+      if (!res.ok)
+        throw new Error(`task-store GET /tasks?${params} → ${res.status}`);
+      return res.json() as Promise<Task[]>;
+    },
+    async update(id: string, fields: Record<string, unknown>): Promise<Task> {
+      const res = await fetch(`${baseUrl}/tasks/${id}`, {
+        method: "PATCH",
+        headers,
+        body: JSON.stringify(fields),
+      });
+      if (!res.ok)
+        throw new Error(`task-store PATCH /tasks/${id} → ${res.status}`);
+      return res.json() as Promise<Task>;
+    },
+  };
+}
+
 // ─── gh CLI helper ────────────────────────────────────────────────────────────
 
 /**
